backend/requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# Google Gemini AI
google-generativeai==0.3.2

# Vector Database
chromadb==1.3.5

# Document Processing
PyPDF2==3.0.1
pypdf==3.17.0
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.3
camelot-py==0.11.0
tabula-py==2.9.0

# Environment Variables
python-dotenv==1.0.0

# Utilities
pydantic==2.5.0
